enable strict deabstraction in diagnostics.swift (SR-8392)
